extract UserSubscriber into it's own module
Breaking the cylce between the project entity and the project repo.
