+++ packages/cli/src/databases/subscribers/index.ts
@@ -1,0 +1,5 @@
+import { UserSubscriber } from './UserSubscriber';
+
+export const subscribers = {
+	UserSubscriber,
+};
